+++ .gitignore
@@ -5,7 +5,7 @@ example_probe
 .idea/
 turbo-go-sdk.iml
 
-pkg/communication/mediation_container_test.go
-pkg/communication/probe_store.go
+pkg/mediationcontainer/mediation_container_test.go
+pkg/mediationcontainer/probe_store.go
 test/
