+++ pkg/mediationcontainer/client_message_builder.go
@@ -1,4 +1,4 @@
-package communication
+package mediationcontainer
 
 import (
 	"github.com/turbonomic/turbo-go-sdk/pkg/proto"
