+++ pkg/mediationcontainer/client_message_builder_test.go
@@ -1,4 +1,4 @@
-package communication
+package mediationcontainer
 
 import (
 	"github.com/stretchr/testify/assert"
