+++ pkg/mediationcontainer/client_websocket_transport.go
@@ -1,4 +1,4 @@
-package communication
+package mediationcontainer
 
 import (
 	"crypto/tls"
