+++ pkg/mediationcontainer/client_websocket_transport_test.go
@@ -1,4 +1,4 @@
-package communication
+package mediationcontainer
 
 import (
 	"testing"
