+++ pkg/mediationcontainer/mediation_container.go
@@ -1,4 +1,4 @@
-package communication
+package mediationcontainer
 
 import (
 	"errors"
