+++ pkg/mediationcontainer/remote_mediation_client.go
@@ -1,4 +1,4 @@
-package communication
+package mediationcontainer
 
 import (
 	"time"
